requirements.txt
neo4j==5.20.0
pandas==2.1.1
requests==2.31.0
beautifulsoup4==4.12.2
scholarly==0.1.1
python-dotenv==1.0.1